+++ Installer&Launcher/InstallerBuildFiles/go.mod
@@ -1,0 +1,3 @@
+module installer.go
+
+go 1.22.5
